requirements.txt
Django==3.1.4
djangorestframework==3.12.2
psycopg2-binary==2.9.7